+++ Andead.SmartHome.Presentation.API/Models/AddWorkflowStepCommand.cs
@@ -3,7 +3,7 @@
     public class AddWorkflowStepCommand
     {
         public string StepName { get; set; }
-        public long WorkflowId { get; set; }
+        public int WorkflowId { get; set; }
         public string ClassName { get; set; }
         public bool IsFirstStep { get; set; }
     }
