+++ Andead.SmartHome.UnitOfWork/Entities/WorkflowAction.cs
@@ -4,7 +4,7 @@
     {
         public string ActionName { get; set; }
 
-        public long WorkflowId { get; set; }
+        public int WorkflowId { get; set; }
 
         public string ClassName { get; set; }
     }
