+++ Andead.SmartHome.UnitOfWork/Extensions/WorkflowActionExtensions.cs
@@ -5,7 +5,7 @@ namespace Andead.SmartHome.UnitOfWork.Extensions
 {
     public static class WorkflowActionExtensions
     {
-        public static IQueryBuilder<WorkflowAction> ByWorkflowId(this IQueryBuilder<WorkflowAction> repository, long workflowId)
+        public static IQueryBuilder<WorkflowAction> ByWorkflowId(this IQueryBuilder<WorkflowAction> repository, int workflowId)
         {
             return repository.Condition(new Specification<WorkflowAction>(x => x.WorkflowId == workflowId));
         }
